Add files via upload
This generates a list of 070 numbers in the UK in the international format. These lists can be imported into any phone system (you may have to hack the resulting files around depending on your setup) This is provided 'as-is' and I do not accept responsibility for any injury, loss of life, or loss of earnings that may result from using this script.